refactor: rename `utils` folder to `common` and move contant to here
