+++ src/common/index.ts
@@ -1,3 +1,6 @@
+import * as C from './constant'
+
+export { C }
 export * from './date-fields'
 export * from './esday-inst'
 export * from './is'
