+++ src/common/is.ts
