+++ src/common/str-utils.ts
